requirements.txt
google-adk
protobuf~=6.31.1